# backend-go/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "基于 Go + MySQL + Redis 的高性能预测系统后端 API，支持体育比赛预测、用户投票、排行榜和实时更新功能\n\n## 功能特性\n- 🏆 体育比赛预测系统\n- 👥 用户注册和认证\n- 🗳️ 预测投票功能\n- 📊 实时排行榜\n- ⚡ WebSocket 实时通信\n- 🚀 高性能缓存策略\n\n## 技术栈\n- **后端**: Go + Gin + GORM\n- **数据库**: MySQL 8.0\n- **缓存**: Redis 6.0\n- **认证**: JWT\n\n## API 版本管理\n当前版本: v1.0，支持向后兼容",
        "title": "预测系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support Team",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/matches/{match_id}/recalculate-points": {
            "post": {
                "description": "管理员使用指定积分规则重新计算某场比赛的所有预测积分",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "积分规则"
                ],
                "summary": "使用指定规则重新计算比赛积分",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "比赛ID",
                        "name": "match_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "积分规则ID（不指定则使用激活规则）",
                        "name": "rule_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/scoring-rules": {
            "post": {
                "description": "管理员创建新的积分规则",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "积分规则"
                ],
                "summary": "创建积分规则",
                "parameters": [
                    {
                        "description": "创建积分规则请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/prediction.CreateScoringRuleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/prediction.ScoringRule"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/scoring-rules/{id}": {
            "put": {
                "description": "管理员更新积分规则信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "积分规则"
                ],
                "summary": "更新积分规则",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "积分规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新积分规则请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/prediction.UpdateScoringRuleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/prediction.ScoringRule"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "管理员删除积分规则（不能删除激活的规则）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "积分规则"
                ],
                "summary": "删除积分规则",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "积分规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/scoring-rules/{id}/activate": {
            "post": {
                "description": "管理员设置当前使用的积分规则",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "积分规则"
                ],
                "summary": "设置激活的积分规则",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "积分规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/scoring-rules": {
            "get": {
                "description": "获取系统中所有的积分规则列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "积分规则"
                ],
                "summary": "获取所有积分规则",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/prediction.ScoringRule"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/scoring-rules/active": {
            "get": {
                "description": "获取当前系统使用的积分规则",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "积分规则"
                ],
                "summary": "获取当前激活的积分规则",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/prediction.ScoringRule"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/scoring-rules/{id}": {
            "get": {
                "description": "获取指定ID的积分规则详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "积分规则"
                ],
                "summary": "获取积分规则详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "积分规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/prediction.ScoringRule"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "prediction.CreateScoringRuleRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "correct_team_correct_score": {
                    "type": "integer",
                    "minimum": 0
                },
                "correct_team_wrong_score": {
                    "type": "integer",
                    "minimum": 0
                },
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                },
                "wrong_team_correct_score": {
                    "type": "integer",
                    "minimum": 0
                },
                "wrong_team_wrong_score": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "prediction.ScoringRule": {
            "type": "object",
            "properties": {
                "correct_team_correct_score": {
                    "description": "预测正确队伍和比分",
                    "type": "integer"
                },
                "correct_team_wrong_score": {
                    "description": "预测正确队伍错误比分",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "wrong_team_correct_score": {
                    "description": "预测错误队伍正确比分",
                    "type": "integer"
                },
                "wrong_team_wrong_score": {
                    "description": "预测错误队伍错误比分",
                    "type": "integer"
                }
            }
        },
...
        "prediction.UpdateScoringRuleRequest": {
            "type": "object",
            "properties": {
                "correct_team_correct_score": {
                    "type": "integer",
                    "minimum": 0
                },
                "correct_team_wrong_score": {
                    "type": "integer",
                    "minimum": 0
                },
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                },
                "wrong_team_correct_score": {
                    "type": "integer",
                    "minimum": 0
                },
                "wrong_team_wrong_score": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "response.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer",
                    "example": 400
                },
                "details": {
                    "description": "Additional error context",
                    "type": "string",
                    "example": "Validation failed"
                },
                "message": {
                    "description": "Primary error message",
                    "type": "string",
                    "example": "Bad request"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response payload (success only)"
                },
                "error": {
                    "description": "Error details (failure only)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.ErrorInfo"
                        }
                    ]
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "description": "Operation success status",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer \" + JWT令牌进行身份验证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}